fix(user-ui): add /api/version
