+++ packages/main-ui/src/main.rs
@@ -63,3 +63,24 @@ fn App() -> Element {
         Router::<Route> {}
     }
 }
+
+#[cfg(feature = "server")]
+mod api {
+    use bdk::prelude::*;
+    use server_fn::codec::{GetUrl, Json};
+
+    #[server(endpoint = "/version", input=GetUrl, output=Json)]
+    pub async fn version() -> Result<String, ServerFnError> {
+        Ok(match option_env!("VERSION") {
+            Some(version) => match option_env!("COMMIT") {
+                Some(commit) => format!("{}-{}", version, commit),
+                None => format!("{}", version),
+            },
+            None => match option_env!("DATE") {
+                Some(date) => date.to_string(),
+                None => "unknown".to_string(),
+            },
+        }
+        .to_string())
+    }
+}
